+++ processor/src/test/resources/bad_input/com/example/component_id_ref/NoReturnModel.java
@@ -1,0 +1,11 @@
+package com.example.component_id_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentIdRef;
+
+@ArezComponent( allowEmpty = true )
+public abstract class NoReturnModel
+{
+  @ComponentIdRef
+  abstract void getId();
+}