+++ processor/src/test/resources/bad_input/com/example/component_id_ref/ParametersModel.java
@@ -1,0 +1,11 @@
+package com.example.component_id_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentIdRef;
+
+@ArezComponent( allowEmpty = true )
+public abstract class ParametersModel
+{
+  @ComponentIdRef
+  abstract int getId( int i );
+}