+++ processor/src/test/resources/expected/com/example/component_id_ref/Arez_NonIntTypeModel.java
@@ -1,0 +1,138 @@
+package com.example.component_id_ref;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.component.ComponentState;
+import arez.component.DisposeNotifier;
+import arez.component.DisposeTrackable;
+import arez.component.Identifiable;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+public final class Arez_NonIntTypeModel extends NonIntTypeModel implements Disposable, Identifiable<String>, DisposeTrackable {
+  private byte $$arezi$$_state;
+
+  @Nullable
+  private final ArezContext $$arezi$$_context;
+
+  private final Component $$arezi$$_component;
+
+  private final DisposeNotifier $$arezi$$_disposeNotifier;
+
+  public Arez_NonIntTypeModel() {
+    super();
+    this.$$arezi$$_context = Arez.areZonesEnabled() ? Arez.context() : null;
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_INITIALIZED;
+    }
+    this.$$arezi$$_component = Arez.areNativeComponentsEnabled() ? $$arezi$$_context().component( "NonIntTypeModel", id(), Arez.areNamesEnabled() ? $$arezi$$_name() : null, () -> $$arezi$$_preDispose() ) : null;
+    this.$$arezi$$_disposeNotifier = new DisposeNotifier();
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_CONSTRUCTED;
+    }
+    if ( Arez.areNativeComponentsEnabled() ) {
+      this.$$arezi$$_component.complete();
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_READY;
+    }
+  }
+
+  final ArezContext $$arezi$$_context() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_context' invoked on uninitialized component of type 'NonIntTypeModel'" );
+    }
+    return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
+  }
+
+  @Override
+  final String getId() {
+    return this.id();
+  }
+
+  @Override
+  @Nonnull
+  public final String getArezId() {
+    return id();
+  }
+
+  String $$arezi$$_name() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> ComponentState.hasBeenInitialized( this.$$arezi$$_state ), () -> "Method named '$$arezi$$_name' invoked on uninitialized component of type 'NonIntTypeModel'" );
+    }
+    return "NonIntTypeModel." + id();
+  }
+
+  private void $$arezi$$_preDispose() {
+    $$arezi$$_disposeNotifier.dispose();
+  }
+
+  @Override
+  @Nonnull
+  public DisposeNotifier getNotifier() {
+    return $$arezi$$_disposeNotifier;
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return ComponentState.isDisposingOrDisposed( this.$$arezi$$_state );
+  }
+
+  @Override
+  public void dispose() {
+    if ( !ComponentState.isDisposingOrDisposed( this.$$arezi$$_state ) ) {
+      this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSING;
+      if ( Arez.areNativeComponentsEnabled() ) {
+        this.$$arezi$$_component.dispose();
+      } else {
+        $$arezi$$_context().safeAction( Arez.areNamesEnabled() ? $$arezi$$_name() + ".dispose" : null, true, false, () -> { {
+          this.$$arezi$$_preDispose();
+        } } );
+      }
+      if ( Arez.shouldCheckApiInvariants() ) {
+        this.$$arezi$$_state = ComponentState.COMPONENT_DISPOSED;
+      }
+    }
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return null != id() ? id().hashCode() : System.identityHashCode( this );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( this == o ) {
+        return true;
+      } else if ( null == o || !(o instanceof Arez_NonIntTypeModel) ) {
+        return false;
+      } else if ( Disposable.isDisposed( this ) != Disposable.isDisposed( o ) ) {
+        return false;
+      } else {
+        final Arez_NonIntTypeModel that = (Arez_NonIntTypeModel) o;
+        return null != id() && id().equals( that.id() );
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arezi$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}