+++ processor/src/test/resources/input/com/example/component_id_ref/CombinedWithComponentIdModel.java
@@ -1,0 +1,18 @@
+package com.example.component_id_ref;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentId;
+import arez.annotations.ComponentIdRef;
+
+@ArezComponent( allowEmpty = true )
+public abstract class CombinedWithComponentIdModel
+{
+  @ComponentId
+  final int id()
+  {
+    return 0;
+  }
+
+  @ComponentIdRef
+  abstract int getId();
+}